Cherry-pick of upstream e424b3879c751a8c0: fixes #4781
Summary:
This is a hand-cherry-picked of a change included in bootstrap upstream (v3.0)
that fixes an infinite loop when double modals are introduced. Sadly, we use
double modals for related videos, as we haven't fleshed out the design for
that experience yet. In the meantime, IE users are crashing hard on this.

Test Plan:
make compiled_and_readable in parent webapp project
load dashboard, open a practice task with related video
click on related video, made sure it opened!

Reviewers: marcos, marcia

Reviewed By: marcos

Differential Revision: http://phabricator.khanacademy.org/D3920