Prevent enforceFocus from occuring against input elements contained inside of a modal
Summary:
Modals attempt to refocus themselves if an element gains focus outside of them.
This is problematic in instances where we have two modals on the same
screen simultaneously (we shouldn't, but unfortunately we do), because it
causes a race condition where one modal steals focus from the other. So far
I have only been able to reproduce in Firefox.

As a fix, in `enforceFocus` I check if the focus event target is inside of
*any* modal, rather than just the modal doing the check.

Test Plan:
. First, we need two modals simultaneously, which we can accomplish by logging in on a restricted domain
  restricted_domains.RestrictedDomain.add('dangerzone.com')
. Then, in Firefox, log into an account on the domain "dangerzone.com"
. See two modals on login - the restricted domain age selector placed in front of the avatar customizer
. See that the select box works as expected

Reviewers: marcia, jlfwong

Reviewed By: jlfwong

Subscribers: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D15868