Change `enforceFocus` to use a synthesized .bootstrap-modal class
Summary:
enforceFocus was causing problems with exercise tasks in IE, where
it would focus the modal div whenever a user tried to click on an
input. This was because the exercise task was missing the appropriate
.modal class.

This commit changes .modal to .boostrap-modal so that the css doesn't
conflict, and makes it so that bootstrap-modal.js adds that class
itself, rather than relying on it existing on the modal element
beforehand.

This fixes the bug in IE so that events originate from the

This wasn't a problem in other browsers because only in IE
is .focus() asynchronous, allowing the modal to get
focused after the input, even though .focus() was called in
the 'focusin' event, before the actual focusing of the input.

Test Plan:
Do "Addition within 20" in Internet Explorer 11

Auditors: jdan, eater