Update bootstrap modal z-index to match recent variable changes
Summary:
In {3aec366dbebe3dcc9e44f3687cda48f11b15067f}, the modal variables we use were updated to place modals above the global nav. Unfortunately, this had a negative affect on some modals which accept their styling from our new `zindexModalBackdropAboveHeader` variable for the backdrop but use the default `zindexModal` variable for the modal! This results in the modal appearing behind the backdrop (and not being accessible to users at all because clicking the backdrop closes the modal).

This diff simply updates the variable the bootstrap modal's styling uses to use our new custom `zindexModalBackdropAboveHeader` variable for the bootstrap modals.

Another way to implement this would be to update `third_party/javascript-khansrc/bootstrap-khansrc/less/variables.less` to use our custom modal variables so this is updated everywhere as there are a few places where we're using the default bootstrap modal variables.

Before:
{F694764}

After:
{F694765}

Other places in the codebase that the `zindexModal` variable is being used:
{F694766}

Asana task: https://app.asana.com/0/257584043292808/327720087612106

Test Plan:
Run `make serve` in your terminal,
navigate to `http://localhost:8080/profile/` in your browser (and login if necessary),
click on "Edit profile" in the top right,
verify that the profile customization modal opens, appears above the gray backdrop, and that you can click on the modal and edit your information.

Reviewers: charlie

Reviewed By: charlie

Differential Revision: https://phabricator.khanacademy.org/D35443