Update bootstrap modal to support ADR-119
Summary:
In ADR #119 we want give #outer-wrapper it's own stacking context
by setting style="position:relative;z-index:0;" on it.  This is so
that elements with z-indexes on them within #outer-wrapper don't
interact with modals, tooltips and other floaty things that wonder-blocks
implements using portals which add themselves to the end of document.body.
Unfortunately, bootstrap modal was add its backdrop to the end of
document.body even though the modal dialog is still within #outer-wrapper.
This was causing the dialog to appear behind the modal after the changes
proposed in ADR #119 were implemented.

I also ran 'make build' which resulted in the majority of the changes
in this diff.  Interestingly the change I made to bootstrap-modal.js fixed
the issue before I ran 'make build' so maybe we use this module in a way
that doesn't requiring rebuilding.

Test Plan: - http://localhost:8081/mission/cc-fourth-grade-math, open a task, see the modal looks and works correctly

Reviewers: john, jsatk, jvoll, miguel

Differential Revision: https://phabricator.khanacademy.org/D49850